Frontend/requirements.txt
streamlit
httpx
pandas
python-dotenv
